frontend: Extract findKubeconfigByClusterName to new file
