+++ frontend/src/components/App/Settings/SettingsCluster.tsx
@@ -34,7 +34,8 @@ import { isElectron } from '../../../helpers/isElectron';
 import { useCluster, useClustersConf } from '../../../lib/k8s';
 import { deleteCluster, parseKubeConfig, renameCluster } from '../../../lib/k8s/apiProxy';
 import { setConfig, setStatelessConfig } from '../../../redux/configSlice';
-import { findKubeconfigByClusterName, updateStatelessClusterKubeconfig } from '../../../stateless';
+import { updateStatelessClusterKubeconfig } from '../../../stateless';
+import { findKubeconfigByClusterName } from '../../../stateless/findKubeconfigByClusterName';
 import ConfirmButton from '../../common/ConfirmButton';
 import ConfirmDialog from '../../common/ConfirmDialog';
 import Empty from '../../common/EmptyContent';
