+++ frontend/src/lib/k8s/api/v1/clusterApi.ts
@@ -19,11 +19,8 @@ import { loadClusterSettings } from '../../../../helpers/clusterSettings';
 import { getHeadlampAPIHeaders } from '../../../../helpers/getHeadlampAPIHeaders';
 import { ConfigState } from '../../../../redux/configSlice';
 import store from '../../../../redux/stores/store';
-import {
-  deleteClusterKubeconfig,
-  findKubeconfigByClusterName,
-  storeStatelessClusterKubeconfig,
-} from '../../../../stateless';
+import { deleteClusterKubeconfig, storeStatelessClusterKubeconfig } from '../../../../stateless';
+import { findKubeconfigByClusterName } from '../../../../stateless/findKubeconfigByClusterName';
 import { getCluster, getSelectedClusters } from '../../../cluster';
 import { ClusterRequest, clusterRequest, post, request } from './clusterRequests';
 import { JSON_HEADERS } from './constants';
